Second round edits, first part
Notes
L71 I’ve removed a chunk of text that was repeating the main text in the tip just below it. I thought that was confusing
L76 – I’ve added a sentence saying what knit is. Otherwise the sentence compared Preview to Knit but it’s the first time that we have introduced Knit.